+++ packages/plurals/__tests__/rules.test.ts
@@ -2,11 +2,11 @@ import { Decimal } from '@phensley/decimal';
 import { pluralRules } from '../src';
 
 test('missing rules', () => {
-  const op = pluralRules.operands(new Decimal('12345.678'));
+  const n = new Decimal('12345.678');
 
-  let cat = pluralRules.cardinal('xyz', op);
+  let cat = pluralRules.cardinal('xyz', n);
   expect(cat).toEqual('other');
 
-  cat = pluralRules.ordinal('xyz', op);
+  cat = pluralRules.ordinal('xyz', n);
   expect(cat).toEqual('other');
 });