+++ packages/plurals/src/index.ts
@@ -1,6 +1,5 @@
 import { cardinalRules, expressions, ordinalRules } from './autogen.rules';
+export { NumberOperands } from './operands';
 import { PluralRules } from './types';
-
-export { NumberOperands as NumberOperands2 } from './operands';
 
 export const pluralRules = new PluralRules(expressions, cardinalRules, ordinalRules);