+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.22.0]
+### Fixed/Changed
+ - Switched plural calculations to use arbitrary precision for all number operand fields.
+ - Extracted all plural samples to use as test cases.
+ - Moved plural operands from the `decimal` package into `plurals`
+ - Fixed sign issue in subtract
+
 ## [0.21.2]
 ### Fixed/Changed
  - Corrected package dependencies for `@phensley/cldr-types`